LayoutTests/fast/dom/DOMException/XPathException-obsolete.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script>
description("Tests that the obsolete XPathException type is not exposed.");

shouldBe("window.XPathException", "undefined");
</script>
</body>
</html>
