Safari-609.1.11.1/LayoutTests/fast/dom/DOMException/XPathException-obsolete-expected.txt

Tests that the obsolete XPathException type is not exposed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.XPathException is undefined
PASS successfullyParsed is true

TEST COMPLETE
